deployment/requirements/base.txt
Django==5.0.4
djangorestframework==3.15.1
celery==5.4.0
django-celery-beat==2.6.0
redis==5.0.4
channels[daphne]==4.1.0
channels_redis==4.2.0